use less compiler flags in gpsp makefile
i had not tested these compiler flags enough and found that some of them were causing random crashes in games like ff6. this should prevent them with minimal performance difference. it also adds the MMAP_JIT_CACHE flag that was added to the raspberry pi flags in the latest commits